Update METAR_2_CSV.bsh
Fill in many fields.  Bring in Location/altitude from NCEI "History" file. (The "accuracy" field is used for validating the formula found the right data... but this field is for radial information when location data is actually from a GPS (where was my vehicle, where was I for body info at this specific time.)